Add XSS-safe-by-default JSON encoders
JSON is HTML-escaped by default, unless the HTTP header `X-HTML-Escape`
is set to 0. By HMTL-escaping by default, we ensure that even clients
which do not properly HTML-escape JSON data before rendering in to
the DOM are protected. However, the HTTP header allows clients to
request JSON that is not HTML-escaped when escaping is undesirable.