Remove unique constraint on identity_comment
The constraint is now applied via a unique index that is specific
to upvotes/downvotes. A separate migration line in ext_migrations
was created to handle such migrations that are associated with
blessed extensions.

In the same manner, other extensions may create migrations, but should
be done with care to avoid conflicts. Making schema changes that cannot
possibly conflict with either upgrades to the core or with other
extensions is recommended.