Set ssl=false for Postgres in Sandstorm
Sandstorm does not preserve UGO permissions when files are packed.
Instead, files become "world-readable", but this is meaningless
inside a single-user container. Unfortunately, Postgres checks
the permissions on /etc/ssl/private/ssl-cert-snakeoil.key before
starting, and refuses to start if the permissions are "incorrect".

Therefore, we avoid permissions issues by simply disabling TLS/SSL,
which is not needed inside the grain anyway.